[ruuvitag] Add a heartbeat to detect missing ruuvitags  (#6006)
* [ruuvitag] Moved classes to internal package
* [ruuvitag] Added heartbeat to monitor the ruuvi tag

This change adds a heartbeat. It will set the thing off line if no signal is received after the heartbeat time frame is exceeded.
